requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt --no-deps
asyncpg==0.30.0
    # via todo-app (pyproject.toml)
fastapi==0.123.5
    # via todo-app (pyproject.toml)
httpx==0.28.1
    # via todo-app (pyproject.toml)
psycopg2-binary==2.9.10
    # via todo-app (pyproject.toml)
pydantic-settings==2.8.1
    # via todo-app (pyproject.toml)
pytest==8.3.5
    # via todo-app (pyproject.toml)
pytest-asyncio==0.24.0
    # via todo-app (pyproject.toml)
sqlalchemy==2.0.44
    # via todo-app (pyproject.toml)
sqlmodel==0.0.27
    # via todo-app (pyproject.toml)
uvicorn==0.33.0
    # via todo-app (pyproject.toml)
